task_manager/requirements.txt
crispy-bootstrap5==0.7
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.2
django-taggit==4.0.0
djangorestframework==3.14.0